Move DNS stub under behavior
