+++ lib/food_geek/accounts/dns.ex
@@ -1,0 +1,3 @@
+defmodule FoodGeek.Accounts.Dns do
+  @callback lookup(binary(), atom(), atom()) :: [{number(), charlist()}]
+end
