+++ lib/food_geek/accounts/dns/inet_res.ex
@@ -1,0 +1,13 @@
+defmodule FoodGeek.Accounts.Dns.InetRes do
+  @behaviour FoodGeek.Accounts.Dns
+
+  require Logger
+
+  def lookup(domain, class, type) do
+    Logger.info("Resoling MX record for #{domain}")
+
+    domain
+    |> to_charlist()
+    |> :inet_res.lookup(class, type)
+  end
+end
